+++ go.mod
@@ -4,5 +4,6 @@ go 1.15
 
 require (
 	encore.dev v0.0.6
+	github.com/satori/go.uuid v1.2.0
 	golang.org/x/tools v0.0.0-20201121010211-780cb80bd7fb // indirect
 )
